refactor: PathValidator의 default 메소드 제거 후 DefaultPathValidator 검증 로직 이동
- default 메소드의 단점 (테스트, DIP)를 고려하여 로직 처리 위치 변경
